[stdlib] Add all swizzle modes to TMA load and multicast
Add all swizzle modes to TMA load and multicast

GPU_ORIG_REV_ID: 009f45488a08cd7bf438a4396881e6c47c2bda33

+++ mojo/stdlib/src/gpu/id.mojo
@@ -110,12 +110,8 @@ fn block_rank_in_cluster() -> UInt32:
     """Gets the unique identifier for the current thread block (CTA) in the
     cluster across all dimensions. Equivalent to `%cluster_ctarank` in CUDA."""
 
-    return UInt(
-        Int(
-            llvm_intrinsic[
-                "llvm.nvvm.read.ptx.sreg.cluster.ctarank",
-                UInt32,
-                has_side_effect=False,
-            ]()
-        )
-    )
+    return llvm_intrinsic[
+        "llvm.nvvm.read.ptx.sreg.cluster.ctarank",
+        UInt32,
+        has_side_effect=False,
+    ]()
